Move "Open file" action from Editor to Application plugin
We want to generalize this action so that it can also be used to
open files in other panes, like the Notebook pane. This commit is
preparation and it still uses the existing implementation of the
action.

As a consequence, the shortcut for "Open file" becomes a global
shortcut and not restricted to the Editor pane.

A side-effect is that the order of the icons in the main toolbar
changes (the "Open file" icon now comes before "New file"). This
will revert back when the "New file" is also moved to the
Application plugin.